requirements.txt
Nuitka==0.8.4
ordered-set==4.1.0
Pillow==9.1.1
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
shiboken6==6.3.1
